added transport controlled recording feature

+++ src/tascar_jackio.cc
@@ -35,7 +35,7 @@
 /**
 \ingroup apptascar
 */
-class jackio_t : public jackc_t {
+class jackio_t : public jackc_transport_t {
 public:
   /**
      \param ifname Input file name
@@ -44,6 +44,8 @@ public:
      \param freewheel Optionally use freewheeling mode
    */
   jackio_t(const std::string& ifname,const std::string& ofname,
+	   const std::vector<std::string>& ports,const std::string& jackname = "jackio",int freewheel = 0,int autoconnect = 0);
+  jackio_t(double start, double duration,const std::string& ofname,
 	   const std::vector<std::string>& ports,const std::string& jackname = "jackio",int freewheel = 0,int autoconnect = 0);
   ~jackio_t();
   /**
@@ -61,13 +63,16 @@ private:
   bool b_quit;
   bool start;
   bool freewheel_;
+  bool use_transport;
+  uint32_t startframe;
+  uint32_t nframes_total;
   std::vector<std::string> p;
-  int process(jack_nframes_t nframes,const std::vector<float*>& inBuffer,const std::vector<float*>& outBuffer);
+  int process(jack_nframes_t nframes,const std::vector<float*>& inBuffer,const std::vector<float*>& outBuffer,uint32_t tp_frame, bool tp_running);
 };
 
 jackio_t::jackio_t(const std::string& ifname,const std::string& ofname,
 		   const std::vector<std::string>& ports, const std::string& jackname,int freewheel,int autoconnect)
-  : jackc_t(jackname),
+  : jackc_transport_t(jackname),
     sf_in(NULL),
     sf_out(NULL),
     buf_in(NULL),
@@ -76,6 +81,9 @@ jackio_t::jackio_t(const std::string& ifname,const std::string& ofname,
     b_quit(false),
     start(false),
     freewheel_(freewheel),
+    use_transport(false),
+    startframe(0),
+    nframes_total(0),
     p(ports)
 {
   if( !(sf_in = sf_open(ifname.c_str(),SFM_READ,&sf_inf_in)) )
@@ -97,6 +105,7 @@ jackio_t::jackio_t(const std::string& ifname,const std::string& ofname,
     if( !(sf_out = sf_open(ofname.c_str(),SFM_WRITE,&sf_inf_out)) )
       throw TASCAR::ErrMsg("unable to open output file");
   }
+  nframes_total = sf_inf_in.frames;
   buf_in = new float[sf_inf_in.channels * sf_inf_in.frames];
   buf_out = new float[sf_inf_out.channels * sf_inf_in.frames];
   memset(buf_out,0,sizeof(float)*sf_inf_out.channels * sf_inf_in.frames);
@@ -113,27 +122,78 @@ jackio_t::jackio_t(const std::string& ifname,const std::string& ofname,
   }
 }
 
+jackio_t::jackio_t(double starttime, double duration,const std::string& ofname,
+		   const std::vector<std::string>& ports, const std::string& jackname,int freewheel,int autoconnect)
+  : jackc_transport_t(jackname),
+    sf_in(NULL),
+    sf_out(NULL),
+    buf_in(NULL),
+    buf_out(NULL),
+    pos(0),
+    b_quit(false),
+    start(false),
+    freewheel_(freewheel),
+    use_transport(true),
+    startframe(get_srate()*starttime),
+    nframes_total(std::max(1u,uint32_t(get_srate()*duration))),
+    p(ports)
+{
+  //if( !(sf_in = sf_open(ifname.c_str(),SFM_READ,&sf_inf_in)) )
+  //  throw TASCAR::ErrMsg("unable to open input file \""+ifname+"\" for reading.");
+  memset(&sf_inf_out,0,sizeof(sf_inf_out));
+  sf_inf_out.samplerate = get_srate();
+  sf_inf_out.channels = std::max(1,(int)(p.size()));
+  sf_inf_out.format = SF_FORMAT_WAV | SF_FORMAT_FLOAT | SF_ENDIAN_FILE;
+  if( autoconnect ){
+    p.clear();
+    p.push_back("system:capture_1");
+  }
+  if( ofname.size() ){
+    if( !(sf_out = sf_open(ofname.c_str(),SFM_WRITE,&sf_inf_out)) )
+      throw TASCAR::ErrMsg("unable to open output file");
+  }
+  //buf_in = new float[1];
+  buf_out = new float[sf_inf_out.channels * nframes_total];
+  memset(buf_out,0,sizeof(float)*sf_inf_out.channels * nframes_total);
+  //memset(buf_in,0,sizeof(float)*sf_inf_in.channels * sf_inf_in.frames);
+  //sf_readf_float(sf_in,buf_in,sf_inf_in.frames);
+  char c_tmp[100];
+  for(unsigned int k=0;k<(unsigned int)sf_inf_out.channels;k++){
+    sprintf(c_tmp,"in_%d",k+1);
+    add_input_port(c_tmp);
+  }
+}
+
 jackio_t::~jackio_t()
 {
   sf_close(sf_in);
   if( sf_out ){
-    sf_writef_float(sf_out,buf_out,sf_inf_in.frames);
+    sf_writef_float(sf_out,buf_out,nframes_total);
     sf_close(sf_out);
   }
 }
 
 
-int jackio_t::process(jack_nframes_t nframes,const std::vector<float*>& inBuffer,const std::vector<float*>& outBuffer)
+int jackio_t::process(jack_nframes_t nframes,const std::vector<float*>& inBuffer,const std::vector<float*>& outBuffer,uint32_t tp_frame, bool tp_running)
 {
+  bool record(start);
+  if( use_transport )
+    record &= tp_running;
+  //DEBUGS(record);
+  //DEBUGS(tp_frame);
+  //DEBUGS(tp_running);
+  //DEBUGS(pos);
+  //DEBUGS(nframes_total);
   for(unsigned int k=0;k<nframes;k++){
-    if( start && (pos < sf_inf_in.frames) ){
-      for(unsigned int ch=0;ch<outBuffer.size();ch++)
-	outBuffer[ch][k] = buf_in[sf_inf_in.channels*pos+ch];
+    if( record && (pos < nframes_total) ){
+      if( buf_in )
+        for(unsigned int ch=0;ch<outBuffer.size();ch++)
+          outBuffer[ch][k] = buf_in[sf_inf_in.channels*pos+ch];
       for(unsigned int ch=0;ch<inBuffer.size();ch++)
 	buf_out[sf_inf_out.channels*pos+ch] = inBuffer[ch][k];
       pos++;
     }else{
-      if( pos >= sf_inf_in.frames)
+      if( pos >= nframes_total)
 	b_quit = true;
       for(unsigned int ch=0;ch<outBuffer.size();ch++)
 	outBuffer[ch][k] = 0;
@@ -145,18 +205,34 @@ int jackio_t::process(jack_nframes_t nframes,const std::vector<float*>& inBuffer
 void jackio_t::run()
 {
   activate();
-  for(unsigned int k=0;k<(unsigned int)sf_inf_in.channels;k++)
-    if( k<p.size())
-      connect_out(k,p[k]);
+  if( !use_transport ){
+    for(unsigned int k=0;k<(unsigned int)sf_inf_in.channels;k++)
+      if( k<p.size())
+        connect_out(k,p[k]);
   for(unsigned int k=0;k<(unsigned int)sf_inf_out.channels;k++)
     if( k+sf_inf_in.channels < p.size() )
       connect_in(k,p[k+sf_inf_in.channels]);
+  }else{
+    for(unsigned int k=0;k<(unsigned int)sf_inf_out.channels;k++)
+      if( k < p.size() )
+        connect_in(k,p[k]);
+  }
   //sleep( 1 );
   if( freewheel_ )
     jack_set_freewheel( jc, 1 );
+  if( use_transport ){
+    tp_stop();
+    tp_locate(startframe);
+  }
   start = true;
+  if( use_transport ){
+    tp_start();
+  }
   while( !b_quit ){
     usleep( 5000 );
+  }
+  if( use_transport ){
+    tp_stop();
   }
   if( freewheel_ )
     jack_set_freewheel( jc, 0 );
@@ -177,7 +253,7 @@ void usage(struct option * opt)
 int main(int argc, char** argv)
 {
   try{
-    const char *options = "fo:chu";
+    const char *options = "fo:chus:d:";
     struct option long_options[] = { 
       { "freewheeling", 0, 0, 'f' },
       { "output-file",  1, 0, 'o' },
@@ -185,6 +261,8 @@ int main(int argc, char** argv)
       { "autoconnect",  0, 0, 'c' },
       { "unlink",       0, 0, 'u' },
       { "help",         0, 0, 'h' },
+      { "start",        1, 0, 's' },
+      { "duration",     1, 0, 'd' },
       { 0, 0, 0, 0 }
     };
     int opt(0);
@@ -195,6 +273,9 @@ int main(int argc, char** argv)
     std::string jackname("tascar_jackio");
     int freewheel(0);
     int autoconnect(0);
+    bool b_use_transport(false);
+    double start(0);
+    double duration(10);
     std::vector<std::string> ports;
     while( (opt = getopt_long(argc, argv, options,
                               long_options, &option_index)) != EOF){
@@ -214,21 +295,36 @@ int main(int argc, char** argv)
       case 'n':
         jackname = optarg;
         break;
+      case 's':
+        start = atof(optarg);
+        b_use_transport = true;
+        break;
+      case 'd':
+        duration = atof(optarg);
+        b_use_transport = true;
+        break;
       case 'h':
         usage(long_options);
         return -1;
       }
     }
-    if( optind < argc )
-      ifname = argv[optind++];
+    if( !b_use_transport ){
+      if( optind < argc )
+        ifname = argv[optind++];
+    }
     while( optind < argc ){
       ports.push_back( argv[optind++] );
     }
-    if( ifname.size() ){
-      jackio_t jio(ifname,ofname,ports,jackname,freewheel,autoconnect);
+    if( !b_use_transport ){
+      if( ifname.size() ){
+        jackio_t jio(ifname,ofname,ports,jackname,freewheel,autoconnect);
+        jio.run();
+        if( b_unlink )
+          unlink(ifname.c_str());
+      }
+    }else{
+      jackio_t jio(start,duration,ofname,ports,jackname,freewheel,autoconnect);
       jio.run();
-      if( b_unlink )
-        unlink(ifname.c_str());
     }
   }
   catch( const std::exception& e ){
